first draft of utils refactoring, under work
