+++ backend-services/configuration/core/configuration.py
@@ -29,12 +29,12 @@ from artemis_utils import RABBITMQ_URI
 from artemis_utils import REDIS_HOST
 from artemis_utils import redis_key
 from artemis_utils import REDIS_PORT
-from artemis_utils import translate_as_set
-from artemis_utils import translate_asn_range
-from artemis_utils import translate_rfc2622
 from artemis_utils import update_aliased_list
-from artemis_utils.rabbitmq_util import create_exchange
-from artemis_utils.rabbitmq_util import create_queue
+from artemis_utils.rabbitmq import create_exchange
+from artemis_utils.rabbitmq import create_queue
+from artemis_utils.translations import translate_as_set
+from artemis_utils.translations import translate_asn_range
+from artemis_utils.translations import translate_rfc2622
 from kombu import Connection
 from kombu import Consumer
 from kombu import Producer
