+++ backend-services/mitigation/core/mitigation.py
@@ -9,8 +9,8 @@ import requests
 import ujson as json
 from artemis_utils import get_logger
 from artemis_utils import RABBITMQ_URI
-from artemis_utils.rabbitmq_util import create_exchange
-from artemis_utils.rabbitmq_util import create_queue
+from artemis_utils.rabbitmq import create_exchange
+from artemis_utils.rabbitmq import create_queue
 from kombu import Connection
 from kombu import Producer
 from kombu import uuid
